refactor: move load fn to lesson-comments module
