+++ src/lib/lesson-comments.ts
@@ -1,4 +1,6 @@
 import axios from 'utils/configured-axios'
+import {getGraphQLClient} from '../utils/configured-graphql-client'
+import getAccessTokenFromCookie from '../utils/get-access-token-from-cookie'
 
 export type Comment = {
   title?: string
@@ -13,4 +15,62 @@ export async function saveCommentForLesson(slug: string, comment: Comment) {
       {comment},
     )
     .then(({data}) => data)
+}
+
+type GraphQLComment = {
+  comment: string
+  commentable_id: number
+  commentable_type: string
+  created_at: string
+  id: number
+  is_commentable_owner: boolean
+  state: string
+  user: {
+    avatar_url: string
+    full_name: string
+    instructor: {
+      first_name: string
+    }
+  }
+}
+
+export async function loadLessonComments(
+  slug: string,
+  token?: string,
+): Promise<GraphQLComment[]> {
+  token = token || getAccessTokenFromCookie()
+  const graphQLClient = getGraphQLClient(token)
+
+  const variables = {
+    slug: slug,
+  }
+
+  const query = /* GraphQL */ `
+    query getLesson($slug: String!) {
+      lesson(slug: $slug) {
+        comments {
+          comment
+          commentable_id
+          commentable_type
+          created_at
+          id
+          is_commentable_owner
+          state
+          user {
+            avatar_url
+            full_name
+            instructor {
+              first_name
+            }
+          }
+        }
+      }
+    }
+  `
+
+  const {
+    lesson: {comments},
+  } = await graphQLClient.request(query, variables)
+
+  return comments
 }
